몸통과 토핑을 그리는 코드 추가
