+++ CodeStarterCamp_Week1/PrintIcecream.swift
@@ -1,8 +1,16 @@
-//
-//  PrintIcecream.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by Sung Chul Yea on 2022/02/23.
-//
+
 
 import Foundation
+
+func printicecream(of length: Int) {
+    for icecream in 1 ... length {
+        print("***********")
+    }
+}
+
+
+func printbar(of length: Int) {
+    for bar in 1 ... length {
+        print("    | |    ")
+    }
+}
