--- PrintPockbyOrder.swift
@@ -1,8 +1,0 @@
-//
-//  PrintPockbyOrder.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by Sung Chul Yea on 2022/02/23.
-//
-
-import Foundation
